+++ crates/qq-core/src/sessions.rs
@@ -2604,7 +2604,7 @@ fn begin_assistant_message(
     store_id: StoreId,
     identity: RunIdentity,
     message_id: MessageId,
-    turn_ordinal: u16,
+    turn_ordinal: u32,
     channel: TextChannel,
     text: &str,
 ) -> Result<Vec<SessionEventEnvelope>, SessionRuntimeError> {
@@ -2949,7 +2949,7 @@ fn apply_steering_message(
     store_id: StoreId,
     identity: RunIdentity,
     message_id: MessageId,
-    turn_ordinal: u16,
+    turn_ordinal: u32,
 ) -> Result<SessionEventEnvelope, SessionRuntimeError> {
     let transaction = store::begin_unit(connection)?;
     let changed = transaction.execute(
@@ -2984,7 +2984,7 @@ fn record_run_interrupted(
     connection: &mut Connection,
     store_id: StoreId,
     identity: RunIdentity,
-    turn_ordinal: u16,
+    turn_ordinal: u32,
 ) -> Result<Vec<SessionEventEnvelope>, SessionRuntimeError> {
     let transaction = store::begin_unit(connection)?;
     let now = now_ms();
@@ -3064,7 +3064,7 @@ fn record_run_output_truncated(
     connection: &mut Connection,
     store_id: StoreId,
     identity: RunIdentity,
-    turn_ordinal: u16,
+    turn_ordinal: u32,
     continuation: u16,
 ) -> Result<SessionEventEnvelope, SessionRuntimeError> {
     let transaction = store::begin_unit(connection)?;
@@ -5825,7 +5825,7 @@ fn load_message(
                 Ok((
                     row.get::<_, String>(0)?,
                     row.get::<_, String>(1)?,
-                    row.get::<_, u16>(2)?,
+                    row.get::<_, u32>(2)?,
                     row.get::<_, String>(3)?,
                     row.get::<_, String>(4)?,
                     row.get::<_, String>(5)?,
@@ -6032,7 +6032,7 @@ fn load_model_context_with_rewrite_status(
     drop(statement);
 
     // Every committed turn for the session's runs, grouped by run.
-    let mut turns: HashMap<String, Vec<(u16, String, bool)>> = HashMap::new();
+    let mut turns: HashMap<String, Vec<(u32, String, bool)>> = HashMap::new();
     let mut statement = transaction.prepare_cached(
         "SELECT t.run_id, t.turn_ordinal, t.assistant_content_json, t.truncated
              FROM model_turns t JOIN runs r ON r.id = t.run_id
@@ -6042,7 +6042,7 @@ fn load_model_context_with_rewrite_status(
     let rows = statement.query_map([&session], |row| {
         Ok((
             row.get::<_, String>(0)?,
-            row.get::<_, u16>(1)?,
+            row.get::<_, u32>(1)?,
             row.get::<_, String>(2)?,
             row.get::<_, bool>(3)?,
         ))
@@ -6090,7 +6090,7 @@ fn load_model_context_with_rewrite_status(
 
     // Applied steering, per run, in the order it was applied. Each carries
     // the ordinal of the turn whose request first included it.
-    let mut steering: HashMap<String, std::collections::VecDeque<(u16, String)>> = HashMap::new();
+    let mut steering: HashMap<String, std::collections::VecDeque<(u32, String)>> = HashMap::new();
     let mut statement = transaction.prepare_cached(
         "SELECT run_id, turn_ordinal, output FROM messages
              WHERE session_id = ?1 AND steering = 1 AND state = 'complete'
@@ -6099,7 +6099,7 @@ fn load_model_context_with_rewrite_status(
     let rows = statement.query_map([&session], |row| {
         Ok((
             row.get::<_, String>(0)?,
-            row.get::<_, u16>(1)?,
+            row.get::<_, u32>(1)?,
             row.get::<_, String>(2)?,
         ))
     })?;
@@ -6431,7 +6431,7 @@ fn search_session_history(
         )?;
         let turns = statement
             .query_map([run_id.to_string()], |row| {
-                Ok((row.get::<_, u16>(0)?, row.get::<_, String>(1)?))
+                Ok((row.get::<_, u32>(0)?, row.get::<_, String>(1)?))
             })?
             .collect::<Result<Vec<_>, _>>()?;
         drop(statement);
@@ -6557,9 +6557,9 @@ fn compaction_instruction(
 /// steering placed immediately before the turn whose request first carried it
 /// and the continuation notice after a truncated turn.
 fn append_run_turns(
-    turns: Vec<(u16, String, bool)>,
+    turns: Vec<(u32, String, bool)>,
     mut recorded: HashMap<String, RecordedResult>,
-    mut steering: std::collections::VecDeque<(u16, String)>,
+    mut steering: std::collections::VecDeque<(u32, String)>,
     context: &mut Vec<Message>,
 ) -> Result<(), SessionRuntimeError> {
     for (turn_ordinal, content_json, truncated) in turns {
@@ -6632,7 +6632,7 @@ fn load_tool_call(
                 Ok((
                     row.get::<_, String>(0)?,
                     row.get::<_, String>(1)?,
-                    row.get::<_, u16>(2)?,
+                    row.get::<_, u32>(2)?,
                     row.get::<_, u16>(3)?,
                     row.get::<_, String>(4)?,
                     row.get::<_, String>(5)?,
@@ -11349,7 +11349,7 @@ mod tests {
                 [],
                 |row| {
                     Ok((
-                        row.get::<_, u16>(0)?,
+                        row.get::<_, u32>(0)?,
                         row.get::<_, String>(1)?,
                         row.get::<_, String>(2)?,
                     ))
@@ -12773,7 +12773,7 @@ mod tests {
         assert!(!truncated);
         assert_eq!(continuations, 0);
         // Historical context replays without a continuation notice.
-        let (ordinal, content_json, truncated): (u16, String, bool) = connection
+        let (ordinal, content_json, truncated): (u32, String, bool) = connection
             .query_row(
                 "SELECT turn_ordinal, assistant_content_json, truncated FROM model_turns
                  WHERE run_id = ?1",
@@ -15343,7 +15343,7 @@ mod tests {
             let turns = statement
                 .query_map([run_id.to_string()], |row| {
                     Ok((
-                        row.get::<_, u16>(0)?,
+                        row.get::<_, u32>(0)?,
                         row.get::<_, String>(1)?,
                         row.get::<_, bool>(2)?,
                     ))
@@ -15360,7 +15360,7 @@ mod tests {
             )?;
             let mut steering = statement
                 .query_map([run_id.to_string()], |row| {
-                    Ok((row.get::<_, u16>(0)?, row.get::<_, String>(1)?))
+                    Ok((row.get::<_, u32>(0)?, row.get::<_, String>(1)?))
                 })?
                 .collect::<Result<std::collections::VecDeque<_>, _>>()?;
             drop(statement);
